Inline _mma helpers back into kernel/005_lower_atom_tile
Fold the MMA fragment-chain lowering helpers (lower_atom_cells + emitters)
from the sibling _mma.py module back into 005_lower_atom_tile.py and delete
_mma.py — the module split no longer earns its indirection now that the Atom
spec rides the Mma op and operands match by SSA name (no tagging machinery to
isolate). Also folds in the stale-docstring fix (011 leaves operand Loads
plain; operands match via the Mma's A/B SSA names, not a Load.role tag).
Kernel byte-identical; full suite green.
